fix(physics): handle pre-step and animated for nodes w. parents
