+++ packages/babylon-lite/src/physics/havok.ts
@@ -23,6 +23,7 @@ import { decomposeMat4 } from "../math/decompose-mat4.js";
 import { transformCoordinatesToRef } from "../math/mat4-transform.js";
 import { addVec3 } from "../math/add-vec3.js";
 import { multiplyQuat } from "../math/multiply-quat.js";
+import { havokTransformToNode, nodeToHavokTransform } from "./havok-transform.js";
 
 // ─── Enums ───────────────────────────────────────────────────────────
 
@@ -430,23 +431,8 @@ function _syncBodyToNode(hknp: any, body: PhysicsBody): void {
         return;
     }
     const t = hknp.HP_Body_GetQTransform(body._hkBody)[1];
-    const pos = t[0]; // [x, y, z]
-    const rot = t[1]; // [x, y, z, w]
     const node = body.node;
-    if (node.parent) {
-        const iwm = invertMat4(node.parent.worldMatrix);
-        if (!iwm) {
-            console.warn("Havok: node world matrix is singular, skipping body→node sync", node);
-            return; // singular world matrix, skip sync
-        }
-        const { rotation } = decomposeMat4(iwm);
-        transformCoordinatesToRef(pos[0], pos[1], pos[2], iwm, node.position);
-        const q = multiplyQuat(rotation, { x: rot[0], y: rot[1], z: rot[2], w: rot[3] });
-        node.rotationQuaternion.set(q.x, q.y, q.z, q.w);
-    } else {
-        node.position.set(pos[0], pos[1], pos[2]);
-        node.rotationQuaternion.set(rot[0], rot[1], rot[2], rot[3]);
-    }
+    havokTransformToNode(t, node);
 }
 
 function _syncNodeToBody(hknp: any, body: PhysicsBody): void {
@@ -454,12 +440,7 @@ function _syncNodeToBody(hknp: any, body: PhysicsBody): void {
         return;
     }
     const node = body.node;
-    const p = node.position;
-    const q = node.rotationQuaternion;
-    hknp.HP_Body_SetQTransform(body._hkBody, [
-        [p.x, p.y, p.z],
-        [q.x, q.y, q.z, q.w],
-    ]);
+    hknp.HP_Body_SetQTransform(body._hkBody, nodeToHavokTransform(node));
 }
 
 // ACTION prestep: instead of snapping the body, set its target transform so Havok derives a
@@ -470,12 +451,7 @@ function _syncNodeToBodyTarget(hknp: any, body: PhysicsBody): void {
         return;
     }
     const node = body.node;
-    const p = node.position;
-    const q = node.rotationQuaternion;
-    hknp.HP_Body_SetTargetQTransform(body._hkBody, [
-        [p.x, p.y, p.z],
-        [q.x, q.y, q.z, q.w],
-    ]);
+    hknp.HP_Body_SetTargetQTransform(body._hkBody, nodeToHavokTransform(node));
 }
 
 // ─── Gravity ─────────────────────────────────────────────────────────
@@ -638,23 +614,7 @@ export function createPhysicsBody(world: PhysicsWorld, node: SceneNode, motionTy
     } else {
         // Add to world first, then set transform (Havok resets transform on add)
         hknp.HP_World_AddBody(hkWorld, hkBody, startsAsleep);
-
-        let p: Vec3;
-        let q: Quat;
-        if (node.parent) {
-            const wm = node.parent.worldMatrix;
-            const { translation, rotation } = decomposeMat4(wm);
-            p = addVec3(translation, node.position);
-            q = multiplyQuat(rotation, node.rotationQuaternion);
-        } else {
-            p = node.position;
-            q = node.rotationQuaternion;
-        }
-
-        hknp.HP_Body_SetQTransform(hkBody, [
-            [p.x, p.y, p.z],
-            [q.x, q.y, q.z, q.w],
-        ]);
+        hknp.HP_Body_SetQTransform(hkBody, nodeToHavokTransform(node));
     }
 
     world._bodies.push(body);
@@ -1431,15 +1391,15 @@ export function setPhysicsBodyMotionType(world: PhysicsWorld, body: PhysicsBody,
  * reads the node before the next physics step stays consistent.
  */
 export function setPhysicsBodyTransform(world: PhysicsWorld, body: PhysicsBody, position: Vec3, rotation: Quat): void {
-    world._hknp.HP_Body_SetQTransform(body._hkBody, [
+    const t = [
         [position.x, position.y, position.z],
         [rotation.x, rotation.y, rotation.z, rotation.w],
-    ]);
+    ] as const;
+    world._hknp.HP_Body_SetQTransform(body._hkBody, t);
     if (world._thin?.count(body) !== undefined) {
         return;
     }
-    body.node.position.set(position.x, position.y, position.z);
-    body.node.rotationQuaternion.set(rotation.x, rotation.y, rotation.z, rotation.w);
+    havokTransformToNode(t, body.node);
 }
 
 // ─── Removal ─────────────────────────────────────────────────────────
